@@ -1,55 +1,64 @@
 {
   "name": "practice",
   "engines": {
-    "node": "16.20.2"
+    "node": "20.9.0"
   },
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.8.1",
-    "@emotion/styled": "^11.8.1",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
     "@fortawesome/fontawesome-svg-core": "^1.3.0",
-    "@fortawesome/free-regular-svg-icons": "^6.0.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/react-fontawesome": "^0.1.17",
-    "@material-ui/core": "^4.12.3",
+    "@fortawesome/free-regular-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.1.19",
+    "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
-    "@mui/icons-material": "^5.6.2",
-    "@mui/material": "^5.4.4",
+    "@mui/icons-material": "^5.15.3",
+    "@mui/material": "^5.15.3",
     "@project-serum/anchor": "^0.23.0",
-    "@solana/buffer-layout": "^4.0.0",
-    "@solana/wallet-adapter-base": "^0.7.1",
-    "@solana/wallet-adapter-material-ui": "^0.13.0",
-    "@solana/wallet-adapter-react": "^0.13.1",
-    "@solana/wallet-adapter-wallets": "^0.11.1",
-    "@solana/web3.js": "^1.39.1",
-    "@testing-library/jest-dom": "^5.14.1",
+    "@solana/buffer-layout": "^4.0.1",
+    "@solana/wallet-adapter-base": "^0.9.23",
+    "@solana/wallet-adapter-material-ui": "^0.16.33",
+    "@solana/wallet-adapter-react": "^0.15.35",
+    "@solana/wallet-adapter-wallets": "^0.19.24",
+    "@solana/web3.js": "^1.87.6",
+    "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "@walletconnect/web3-provider": "^1.7.7",
+    "@types/react-dom": "^18.2.18",
+    "@walletconnect/web3-provider": "^1.8.0",
+    "assert": "^2.1.0",
     "bignumber.js": "^9.1.2",
-    "bootstrap": "^5.1.3",
-    "notistack": "^2.0.3",
-    "react": "^17.0.2",
+    "bootstrap": "^5.3.2",
+    "browserify-zlib": "^0.2.0",
+    "buffer": "^6.0.3",
+    "crypto-browserify": "^3.12.0",
+    "https-browserify": "^1.0.0",
+    "notistack": "^2.0.8",
+    "react": "^18.2.0",
     "react-app-rewired": "^2.2.1",
-    "react-bootstrap": "^2.1.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "4.0.3",
+    "react-bootstrap": "^2.9.2",
+    "react-dom": "^18.2.0",
+    "react-redux": "^7.2.9",
+    "react-router-dom": "^6.21.1",
+    "react-scripts": "^5.0.1",
     "react-toastify": "^8.2.0",
-    "react-tsparticles": "^2.0.6",
-    "redux-thunk": "^2.4.1",
-    "typescript": "^4.6.3",
+    "react-tsparticles": "^2.12.2",
+    "redux-thunk": "^2.4.2",
+    "stream-browserify": "^3.0.0",
+    "stream-http": "^3.2.0",
+    "typescript": "^4.9.5",
+    "url": "^0.11.3",
     "web-vitals": "^1.1.2",
-    "web3": "^1.7.0",
-    "web3modal": "^1.9.7"
+    "web3": "^4.3.0",
+    "web3modal": "^1.9.12"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-script eject"
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
+    "eject": "react-app-rewired eject"
   },
   "eslintConfig": {
     "extends": [
@@ -57,12 +66,12 @@
       "react-app/jest"
     ]
   },
+  "devDependencies": {
+    "nodemon": "^3.0.2"
+  },
   "browserslist": [
     ">0.2%",
     "not dead",
     "not op_mini all"
-  ],
-  "devDependencies": {
-    "nodemon": "^2.0.15"
-  }
+  ]
 }
