@@ -1,5 +1,5 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import React, { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 import CssBaseline from "@mui/material/CssBaseline";
 
 import "./index.css";
@@ -9,16 +9,17 @@ import ThemeProvider from "./Theme";
 
 import { ContractProvider } from "./providers/ContractProvider";
 
-ReactDOM.render(
+createRoot(document.getElementById("root")!).render(
   <>
-    <CssBaseline />
-    <ThemeProvider>
-      <ContractProvider>
-        <App />
-      </ContractProvider>
-    </ThemeProvider>
-  </>,
-  document.getElementById("root")
+    <StrictMode>
+      <CssBaseline />
+      <ThemeProvider>
+        <ContractProvider>
+          <App />
+        </ContractProvider>
+      </ThemeProvider>
+    </StrictMode>
+  </>
 );
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
